ELT/requirements.txt
pandas
bs4
sqlalchemy 
psycopg2-binary
python-dotenv
requests